Add support for building a riscv32 GCC toolchain for CentOS 6
This adds the common build script for building a multilib enabled
riscv32 GNU toolchain and Jenkinsfile, Dockerfile and support files
nedded to checkout, build, package and test such a toolchain for
CentOS 6.

The common scripts will be used for other operating system targets,
leaving just the Jenkinsfile and Dockerfile to be duplicated and
modified as required.