Add Windows build
This is based on the existing CentOS builds, but instead runs on a
Windows machine with a pre-installed MinGW 64-bit installation.

To keep the builds as common as possible, it uses the same common stages
as the Linux builds, instead invoking MSYS bash in MINGW64 mode for each
build stage. The build will also automatically copy required MinGW DLLs
as part of the generated zip file.