Add support for building Clang toolchain for CentOS 7 and 8
This combines clang + newlib + compiler-rt with GNU binutils + gdb to
produce a full toolchain that can run on CentOS 7 and 8. This currently
supports the default bare metal targets of rvXLENimc as the build
target; this will be expanded as better multilib support is upstream.

When put next to the equivalent GCC build its multilibs can be used to
support a wider array of configurations.