Add master trigger script and pull sources of shared mirror
This introduces a couple of changes to the build system:

- Adds a master job which will trigger all the variants to be built
- Has the master push to a local server a copy of the four largest
  repositories used by the builds (binutils-gdb/gcc/newlib-cygwin/
  llvm-project).

This has the effect of keeping all the builds running off the same
commits, even if a builder is not available for a while. Additionally,
it reduces the load we put on upstream repositories since a download of
the top of tree commit will be done 20 times with the current GCC and
LLVM combinations for riscv32/or1k.

describe-build.sh will indicate that these mirrors have been used, but
these are also public so can be verified against upstream sources
(as a pure mirror all hashes will be the same, which makes such a
comparison easy).

Note: the release builders are deliberately not changed in this way,
they will continue to use upstream sources (which is less of an issue
since these jobs run far, far less frequently).