Move the gcc sources to "riscv-gcc"
The binutils-gdb source used in corev-* requires gcc to be checked out
in a directory called "riscv-gcc" since some source files include files
found in this repository. This changes the Jenkinsfiles for all these
builds to check out source in this directory, and the build script to
use this new directory.

Note: This change means that only the corev-* files are valid in this
branch, it should be reverted when binutils-gdb no longer includes files
from the GCC sources.