Add --with-system-zlib to riscv32 GCC configure arguments
It appears that for the gcc used in the corev-* builds, building the
zlib inside the gcc repository does not work when --enable-multilib is
provided; it tries to configure for multilibs on the host and then
the configure fails. As such, switch to using the system zlib for these
builds.