Sign and notarize macOS packages
This adds scripts to code sign and notarize a macOS build directory
using credentials specified in the builder.

Instead of generating a .tar.gz file, the outputs are now a .zip file,
and a .dmg containing the build tools. Due to the limitations in
notarization, the .dmg will have a ticket stapled such that Gatekeeper
will allow the tools to be used offline; this will become the preferred
distribution method of future toolchains.