Set custom CFLAGS/CXXFLAGS for building CGEN simulator
The CGEN simulator uses implicitly defined functions, even though this
is invalid in C99 and above. Newer versions of the macOS Xcode Clang
compiler enable -Werror=implicit-function-declaration by default* which
stops this build from succeeding.

This adds the flag -Wno-error=implicit-function-declaration to the
affected stages, such that the build can now succeed. In the future,
when the simulator is warning clean, this change can be removed.

* https://developer.apple.com/documentation/xcode-release-notes/xcode-12-release-notes