riscv32-gcc: Use --with-multilib-generator option
As of GCC commit c1e6691245ca2f1f329549f323f67afe32bcb97a, the configure
option --with-multilib-generator allows the set of multilibs to be set
at configure time rather than with an ahead of time change to the GCC
source. This switches the builds to use this newer scheme.

Additionally, since the multilib generator script is called via an
unnumbered python, install python2 on CentOS8 and Ubuntu and set that as
the default python to match the other build configurations.