riscv32-gcc: Allow triple to be overridden by defining TRIPLE
This allows the triple and default architecture/ABI to be changed from
their defaults using the TRIPLE/DEFAULTARCH/DEFAULTABI environment
variables.

Note this change is a placeholder to enable functionality whilst
awaiting migration to updated build scripts. This is currently on
hold for a corresponding change to the various Jenkinsfiles (switching
from environment variable control to parameter control).