binutils: Remove gas/doc/.dirstamp
The binutils-2_40 tag contains an accidently committed dirstamp file,
which causes builds to fail when using this tag. As we may wish to use
binutils 2.40, and our build flow involved checking out the repository,
add a stage to the build scripts to remove this file before attempting
to build binutils.