riscv32-clang: Move to explicit distribution component list
In order to ship LLVMgold as part of the installation, move from using
the default set of components installed by specifying
LLVM_INSTALL_TOOLCHAIN_ONLY, and explicitly specify the set of relevant
components via LLVM_DISTRIBUTION_COMPONENTS and install those. This
component list was built by reviewing LLVM's current
LLVM_TOOLCHAIN_TOOLS list, and keeping those that are likely to be
useful in this context.

Alongside this, since the aim is to enable LTO, this also starts
building and installing LLD as part of the toolchain.