riscv32-clang-win64: Disable LLVM threads
We are experiencing an issue using winpthreads which can cause LLVM to
deadlock. To work around this for the Windows builds we disable LLVM's
threading.

This also fixes an issue in the build script where EXTRA_LLVM_OPTS was
being overridden unconditionally, and updates its name to be consistent
with the other project-specific variables.