riscv32-clang: Move to a baremetal clang config
This replaces the pre-existing Clang build configuration to one using
the LLVM Clang Baremetal configuration and its associated multilibs. In
addition this removes binutils and GDB instead relying purely on LLVM
tooling.

Since the old configuration may be useful to some users, the old build
script is kept but is unused. Instead this adds a
'riscv32-clang-baremetal' script which builds the new tools.