riscv-embedded: Add new combined GCC+LLVM 'embedded' builds
This adds a new build configuration for a combined GCC and Clang/LLVM
toolchain which is being labelled 'riscv-embedded'. This uses the
existing GCC configuration to build a toolchain, and adds just the
Clang/LLVM build from the riscv32-clang configuration (i.e. newlib is
built by GCC, and compiler-rt is not included).

Under the hood there is a new 'riscv32-clang-only' build script that
adds the LLVM configuration. In the future this may get merged into
one of the existing scripts with a new mode, but whilst the build
configuration is relatively simple, the configurability would make
the script more complicated so leaving the configuration duplicated is
more maintainable for now.