riscv32-gcc: Add -Wno-error=incompatible-pointer-types to simulator build
The CGEN simulator no longer builds with newer C/C++ tools where the
size of the time types do not match between target and host. For now,
disable the conversion of this warning into an error to allow the
simualtor to work. We do not use any time functions in testing, so this
has no meaningful change to actual testing.